==> sampled articles/code/1_wrangle_issns.Rmd <==
---
title: "Wrangle ISSNs"
author: "Ian Hussey & Jamie Cummins"
date: "`r format(Sys.time(), '%d %B %Y')`"
output:
  html_document:
    code_folding: hide
    highlight: haddock
    theme: flatly
    toc: yes
    toc_float: yes
---

```{r, include = FALSE}
knitr::opts_chunk$set(echo = TRUE, eval = TRUE, include = TRUE, warning = FALSE, message = FALSE)
```

# Depencencies

```{r}

library(dplyr)
library(readr)
library(tidyr)
library(stringr)
library(janitor)

```

# Processing

```{r}

dat <- bind_rows(
  read_csv("../data/journals/JCR_psych_1.csv", skip = 2) |> 
    mutate(source = "psych1"),
  read_csv("../data/journals/JCR_psych_2.csv", skip = 2) |> 
    mutate(source = "psych2"),
  read_csv("../data/journals/manual_additions.csv") |> 
    mutate(source = "manual additions")
) |>
  # wrangling
  janitor::clean_names() |>
  select(-notes) |>
  mutate(journal_name = tolower(journal_name)) |>
  distinct(journal_name, issn, e_issn, .keep_all = TRUE) |>
  mutate(issn = ifelse(issn == "N/A", NA, issn),
         e_issn = ifelse(e_issn == "N/A", NA, e_issn)) |>
  # exclude all journals with "review" or current opinion in the title
  filter(!str_detect(journal_name, "review")) |>
  filter(!str_detect(journal_name, "current opinion")) |>
  # exclude specific well known non-empirical (or mostly non-empirical) journals
  filter(!journal_name %in% c("psychological bulletin", 
                              "psychological science in the public interest", 
                              "perspectives on psychological science",
                              "behavioral and brain sciences",
                              "current directions in psychological science",
                              "trends in cognitive sciences", 
                              "psychological inquiry",
                              "american psychologist", # mostly non-empirical
                              "educational psychologist", # mostly non-empirical
                              "industrial and organizational psychology-perspectives on science and practice", # mostly non-empirical
                              "clinical psychology-science and practice")) |>
  # add IF rank
  mutate(impact_factor_rank = row_number()) |>
  # reshape
  pivot_longer(cols = c("issn", "e_issn"),
               names_to = "ISSN_type", 
               values_to = "ISSN") |>
  drop_na() |>
  distinct(ISSN, .keep_all = TRUE)

write_csv(dat, "../data/journals/processed_journal_issns.csv")

```

# Session info

```{r}

sessionInfo()

```
